uncompressed: pulled PixelInterleaveDecoder into its own file
